+++ src/main/java/com/willshex/blogwt/server/service/invoice/invoice/IInvoiceService.java
@@ -64,8 +64,8 @@ public interface IInvoiceService extends IService {
 	 * @param sortDirection
 	 * @return
 	 */
-	public List<Invoice> getVendorInvoices (Vendor vendor, Long start,
-			Long count, InvoiceSortType sortBy,
+	public List<Invoice> getVendorInvoices (Vendor vendor, Integer start,
+			Integer count, InvoiceSortType sortBy,
 			SortDirectionType sortDirection);
 
 	/**
@@ -77,7 +77,7 @@ public interface IInvoiceService extends IService {
 	 * @return
 	 */
 	public List<Invoice> getVendorOutstandingInvoices (Vendor vendor,
-			Long start, Long count, InvoiceSortType sortBy,
+			Integer start, Integer count, InvoiceSortType sortBy,
 			SortDirectionType sortDirection);
 
 	/**
